requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
psycopg2-binary==2.9.1
Flask-Redis==0.4.0
python-dotenv==0.19.1
werkzeug==2.0.2
SQLAlchemy==1.3.24